docs: split pre-processor from extension
