+++ .docs/extensions/pyproject.toml
@@ -2,3 +2,4 @@
 typeCheckingMode = "strict"
 reportPrivateUsage = false
 reportUnnecessaryTypeIgnoreComment = "error"
+extraPaths = [".."]
